fix(engine): keep surface-tile planner's MVP in f64, narrow only at GPU upload
composeBodyMvp composed proj*view*model in f64 but narrowed to f32 before
returning. That's fine for GPU-drawing callers, but prepareEarthFrame also
fed the same matrix straight into cutSurfaceTiles, a CPU-side walk that
evaluates mvp*p at ground points metres from the camera. There the w-row
cancels its own radiusMpc-scale terms down to ~1e-21 at low altitude, so
f32-rounding a matrix element beforehand already injects a ~1% relative
error — enough to corrupt the walk's bbox-cull test and silently drop
correctly-visible ancestor subtrees (verified against the live capture in
cut-replay-exact-report.md, where the cut collapsed from dozens of tiles to 6).

composeBodyMvp now returns the raw Float64Array; every GPU-drawing caller
narrows its own copy via narrowMat4 at its upload site (9 call sites:
drawFlooredSpherePick, cloudShellLayer, ringsLayer, atmosphereShellLayer,
texturedBodiesLayer, starSpheresLayer, planetsLayer, fieldStarSphereLayer,
earthLayer.draw). prepareEarthFrame's mvpLocal and cutSurfaceTiles'
viewProjLocal stay Float64Array end-to-end.

New failing-first test in cutSurfaceTiles.test.ts reproduces the collapse
with a generic (non-axis-aligned) low-altitude pose — an axis-aligned nadir
pose has enough incidental symmetry that no ancestor bbox ever straddles
the frustum edge, so the bug needed a properly generic camera pose to show up.

+++ src/utils/scene/cutSurfaceTiles.ts
@@ -34,8 +34,11 @@ export function cutSurfaceTiles(input: {
    *  unit sphere). */
   readonly camPosLocal: Readonly<Vec3>;
   /** View-projection for that frame, column-major. Only x/y extent is read,
-   *  so the depth convention doesn't matter here. */
-  readonly viewProjLocal: Float32Array;
+   *  so the depth convention doesn't matter here. MUST stay `Float64Array` —
+   *  at low altitude the `w`-row cancels its own large terms down to ~1e-21,
+   *  so f32-rounding an element beforehand injects a ~1% relative error that
+   *  corrupts the bbox-cull test below (see `composeBodyMvp`'s header). */
+  readonly viewProjLocal: Float64Array;
   readonly viewportPx: Readonly<Vec2>;
   /** The level the whole-globe base texture already delivers — the walk's floor. */
   readonly baseLevel: number;
